Add base application
This isn't anything useful yet. I am still working on the porting
the parameters from ruby over to python and I am trying to do it
correctly this time.  It doesn't look like the ArgumentParser
properly handles argument groups nested in mutually exclusive
argument groups.  Might need to ask the community about this one
to figure out what I am doing wrong or if that just isn't supported.
Also, I should double check the PEP for the ArgumentParser so that
I can make sure that what I am trying to do is part of the expected
usage. If not, propose a new PEP to cover that.