Add methods to retrieve the config and engine
getEngine was added to create an appropriate SQLAlchemy engine

getConfig was added to return the active configuration file

Modifications were made to the argument list to handle a config
file not in the default path.

There is some code in this file that is garbage. Will update again
when I get a few more minutes.  Real functionality should ensue
soon.