Add license and finalize arguments
Added the license that will govern this code base.  I have decided
on Beerware. There is no guarantee that this is what the license
will be for good, but currently I don't care very much since this
script doesn't do very much, I'll just accept a beer if you want
to buy me one. :-)

I finalized the command-line arguments for the script.  There won't
be any changes to this for atleast a complete version. I am pretty
sure that if there is another version in the future, that version
will actually have a low level packet filter code associated with
it.  I like the idea if it being easy to manage a simple firewall
from a normal shell.