Start adding in database connections
I am just a couple commits from production ready.  I might change
some of the code layouts so that its more object oriented.  This
does however create a SQLite db and work nicely.

The change to the config file is just temporary and will be
reverted. This was for testing.