Move Rule Specific Functions into Rule Class
Prior to this commit, some important Rule functions were
separate from the Rule class. I went ahead and refactored
them into the class where they really belong (well for now).

I do suspect that there will be another major refactor in the
future, but for now, I think code position will be pretty static.

Most of my code refactors are going to be for elegance of coding
and probably not for anything else. :-)